Moving stuff back in board component
